requirements.txt
pytest
faiss-cpu
prompt_toolkit
typer
pyyaml
pytest-cov